If there are no routes, you can create them in the database using the button
