+++ app/src/main/res/layout/activity_main.xml
@@ -12,16 +12,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <TextView
-        android:id="@+id/no_trips_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Brak wycieczek w bazie danych"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <Button
         android:id="@+id/exit_button"
         android:layout_width="wrap_content"
