+++ app/src/main/res/layout/fragment_no_trips.xml
@@ -1,13 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:gravity="center">
+    tools:context=".ui.mainactivity.NoTripsFragment">
 
     <TextView
         android:id="@+id/no_trips_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Brak wycieczek w bazie danych"
-        android:textSize="18sp" />
-</FrameLayout>
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/fill_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Uzupełnij"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/no_trips_text" />
+</androidx.constraintlayout.widget.ConstraintLayout>
